US121812 formatAvailabilityDateString in util
