+++ d2l-sequence-navigator/d2l-lesson-header.js
@@ -4,7 +4,7 @@
 */
 import { CompletionStatusMixin } from '../mixins/completion-status-mixin.js';
 import { ASVFocusWithinMixin } from '../mixins/asv-focus-within-mixin.js';
-import { createDateFromObj } from '../util/util.js';
+import { formatAvailabilityDateString } from '../util/util.js';
 import 'd2l-offscreen/d2l-offscreen.js';
 import '@brightspace-ui/core/components/colors/colors.js';
 import 'd2l-typography/d2l-typography.js';
@@ -419,7 +419,7 @@ class D2LLessonHeader extends ASVFocusWithinMixin(CompletionStatusMixin()) {
 			return;
 		}
 		const { startDate, endDate } = properties;
-		return this._formatAvailabilityDateString(startDate, endDate);
+		return formatAvailabilityDateString(this.localize, startDate, endDate);
 	}
 
 	_getAvailabilityDateTooltip(properties) {
@@ -427,44 +427,7 @@ class D2LLessonHeader extends ASVFocusWithinMixin(CompletionStatusMixin()) {
 			return;
 		}
 		const { startDate, endDate } = properties;
-		return this._formatAvailabilityDateString(startDate, endDate, true);
-	}
-
-	_formatAvailabilityDateString(startDateObj, endDateObj, forTooltip) {
-		const tooltipText = forTooltip ? '.tooltip' : '';
-		const format = forTooltip ? 'medium' : 'shortMonthDay';
-		const formatFunction = forTooltip ? formatDateTime : formatDate;
-
-		const startDate = createDateFromObj(startDateObj);
-		const endDate = createDateFromObj(endDateObj);
-
-		if (startDate && endDate) {
-			return this.localize(
-				`sequenceNavigator.dateRange${tooltipText}`,
-				'startDate',
-				formatFunction(startDate, { format }),
-				'endDate',
-				formatFunction(endDate, { format })
-			);
-		}
-
-		if (startDate) {
-			return this.localize(
-				`sequenceNavigator.starts${tooltipText}`,
-				'startDate',
-				formatFunction(startDate, { format })
-			);
-		}
-
-		if (endDate) {
-			return this.localize(
-				`sequenceNavigator.ends${tooltipText}`,
-				'endDate',
-				formatFunction(endDate, { format })
-			);
-		}
-
-		return '';
+		return formatAvailabilityDateString(this.localize, startDate, endDate, true);
 	}
 
 }
