+++ util/util.js
@@ -1,3 +1,5 @@
+import { formatDate, formatDateTime } from '@brightspace-ui/intl/lib/dateTime.js';
+
 function isMobile() {
 	return /iP[ao]d|iPhone|Android|Windows (?:Phone|CE)|PlayBook|BlackBerry|Vodafone|Mobile/.test(window.navigator.userAgent);
 }
@@ -66,11 +68,48 @@ function createDateFromObj(dateObj) {
 	);
 }
 
+function formatAvailabilityDateString(localize, startDateObj, endDateObj, forTooltip) {
+	const tooltipText = forTooltip ? '.tooltip' : '';
+	const format = forTooltip ? 'medium' : 'shortMonthDay';
+	const formatFunction = forTooltip ? formatDateTime : formatDate;
+
+	const startDate = createDateFromObj(startDateObj);
+	const endDate = createDateFromObj(endDateObj);
+
+	if (startDate && endDate) {
+		return localize(
+			`sequenceNavigator.dateRange${tooltipText}`,
+			'startDate',
+			formatFunction(startDate, { format }),
+			'endDate',
+			formatFunction(endDate, { format })
+		);
+	}
+
+	if (startDate) {
+		return localize(
+			`sequenceNavigator.starts${tooltipText}`,
+			'startDate',
+			formatFunction(startDate, { format })
+		);
+	}
+
+	if (endDate) {
+		return localize(
+			`sequenceNavigator.ends${tooltipText}`,
+			'endDate',
+			formatFunction(endDate, { format })
+		);
+	}
+
+	return '';
+}
+
 export {
 	isMobile,
 	isIOS,
 	isSafari,
 	parseUrlQueryParameters,
 	redirectCS,
-	createDateFromObj
+	formatAvailabilityDateString
 };
